+++ pw/plugins/__init__.py
@@ -1,0 +1,2 @@
+from pw.plugins.html_plugin import HTMLPlugin  # noqa
+from pw.plugins.json_plugin import JSONPlugin  # noqa