+++ CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2026-01-02
+
+### Added
+
+* Refactoring and implement import/export functionalities for different
+  file formats as "Plugins". These Plugins depend only on a callable
+  session factory.
+
 ## [1.1.0] - 2026-01-01
 
 ### Added